show lsblk information when mouse hovering cpu widget

+++ .config/awesome/autostart.pl
@@ -18,7 +18,7 @@ my %programmes = (
 );
 foreach my $key(keys %programmes){
 	if(lc($key) eq 'launchy'){
-		system('pkill -9 '.$key);
+		system('pkill -9 '.$key) if `pgrep $key|wc -l` == 1;
 	}
 	system($programmes{$key}) if `pgrep $key|wc -l` == 0;
 }
